Feature: Add language detection for .sample files in Git Hook Scripts
Added a new condition to the 'detect' function in '/Users/evgeniievstafev/PycharmProjects/pr/github-comment-bot/project/lngdetector/detect.py'. If the file path ends with '.sample' and contains '.git/hooks/', it now returns 'Git Hook Script' as the language. This is to provide accurate language detection for sample Git hook scripts.